libbpf-rs: Move map definitions into separate file
